+++ src/main/java/dev/hytalemodding/config/QualityConfig.java
@@ -179,8 +179,10 @@ public class QualityConfig {
             "Weapon_Arrow",
             "Weapon_Dart",
             "Weapon_Spellbook",
+            "Weapon_Grenade_Frag",
             "Tool_Feedbag",
-            "Tool_Watering_Can"
+            "Tool_Watering_Can",
+            "Tool_Shears_Basic"
     };
 
     public QualityConfig() {}